talki-backend/requirements.txt
fastapi             # Framework làm Backend (Nhanh, hiện đại)
uvicorn             # Cái máy chủy để chạy code FastAPI
sqlalchemy[asyncio]   # Thư viện để code Python nói chuyện với Database
asyncpg             # Driver giúp kết nối Postgres siêu nhanh (bất đồng bộ)
httpx               # Async HTTP client – dùng để gọi Supabase Auth API
python-multipart    # Để nhận file âm thanh upload lên
google-genai        # Thư viện chính chủ để gọi Gemini (API mới)
python-dotenv       # Để đọc file cấu hình .env
pydantic-settings   # Quản lý config từ .env
pyjwt               # Xác thực JWT token từ Supabase
# Voice features:
google-cloud-speech       # STT tiếng Việt
google-cloud-texttospeech # TTS – tổng hợp giọng Boss
psycopg2-binary             # Driver giúp kết nối Postgres siêu nhanh (bất đồng bộ)
cloudinary                 # Media hosting (Cloudinary) - official SDK
